+++ src/lib/wordpress-provider/types.ts
@@ -1,0 +1,3 @@
+export interface WordPressProvider {
+	setupWordPressSite( path: string, wpVersion?: string ): Promise< boolean >;
+}
