+++ src/site-server.ts
@@ -1,11 +1,10 @@
-import { net } from 'electron';
 import fs from 'fs';
 import nodePath from 'path';
 import * as Sentry from '@sentry/electron/main';
 import fsExtra from 'fs-extra';
 import { parse } from 'shell-quote';
 import { deleteSiteCertificate, generateSiteCertificate } from 'src/lib/certificate-manager';
-import { pathExists, recursiveCopyDirectory, isEmptyDir } from 'src/lib/fs-utils';
+import { pathExists } from 'src/lib/fs-utils';
 import { getSiteUrl } from 'src/lib/get-site-url';
 import { addDomainToHosts, removeDomainFromHosts, updateDomainInHosts } from 'src/lib/hosts-file';
 import { decodePassword } from 'src/lib/passwords';
@@ -15,15 +14,13 @@ import { startProxyServer } from 'src/lib/proxy-server';
 import { getPreferredSiteLanguage } from 'src/lib/site-language';
 import SiteServerProcess from 'src/lib/site-server-process';
 import { updateSiteUrl } from 'src/lib/update-site-url';
+import { getWordPressProvider } from 'src/lib/wordpress-provider';
 import WpCliProcess, { MessageCanceled, WpCliResult } from 'src/lib/wp-cli-process';
-import { purgeWpConfig, verifyWordPressChecksums } from 'src/lib/wp-versions';
 import { createScreenshotWindow } from 'src/screenshot-window';
-import { copyBundledLatestWPVersion } from 'src/setup-wp-server-files';
 import { getSiteThumbnailPath } from 'src/storage/paths';
 import { getWpNowConfig } from 'vendor/wp-now/src';
 import { WPNowMode } from 'vendor/wp-now/src/config';
 import { DEFAULT_PHP_VERSION, SQLITE_FILENAME } from 'vendor/wp-now/src/constants';
-import { getWordPressVersionPath, downloadWordPress } from 'vendor/wp-now/src/download';
 
 const servers = new Map< string, SiteServer >();
 const deletedServers: string[] = [];
@@ -32,41 +29,8 @@ export async function createSiteWorkingDirectory(
 	path: string,
 	wpVersion = 'latest'
 ): Promise< boolean > {
-	try {
-		if ( ( await pathExists( path ) ) && ! ( await isEmptyDir( path ) ) ) {
-			// We can only create into a clean directory
-			return false;
-		}
-
-		const wpVersionPath = getWordPressVersionPath( wpVersion );
-		const wpVersionExists = await pathExists( wpVersionPath );
-
-		if ( ! wpVersionExists ) {
-			if ( net.isOnline() ) {
-				try {
-					await downloadWordPress( wpVersion, { overwrite: false } );
-				} catch ( error ) {
-					console.error( `Failed to download WordPress version ${ wpVersion }:`, error );
-					throw new Error(
-						`Failed to download WordPress version ${ wpVersion }. Please try a different version.`
-					);
-				}
-			} else if ( wpVersion === 'latest' ) {
-				await copyBundledLatestWPVersion();
-			} else {
-				return false;
-			}
-		}
-
-		await verifyWordPressChecksums( wpVersion );
-		await purgeWpConfig( wpVersion );
-		await recursiveCopyDirectory( getWordPressVersionPath( wpVersion ), path );
-
-		return true;
-	} catch ( error ) {
-		console.error( 'Error in createSiteWorkingDirectory:', error );
-		throw error;
-	}
+	const provider = getWordPressProvider();
+	return provider.setupWordPressSite( path, wpVersion );
 }
 
 export async function stopAllServersOnQuit() {
